feat(M20DS-55): refactor MISingleFileInput and SingleFileInput components, move shared types and utility functions to singleFileInput.ts
