+++ src/components/MISingleFileInput/utils.ts
@@ -1,6 +1,6 @@
 import { alpha, SxProps } from '@mui/system';
 import { theme } from '@theme';
-import { UploadStatus } from './MISingleFileInput';
+import { UploadStatus } from 'types/singleFileInput';
 
 function getDashedBorderStyle(color: string): SxProps {
   return {
@@ -17,45 +17,12 @@ function getDashedBorderStyle(color: string): SxProps {
 }
 
 /**
- * Truncate file name string if it is longer than 30 characters.
- * Keeps the file extension.
- *
- * @param fileName
- * @returns truncated file name
- */
-export const truncateFileName = (fileName: string) => {
-  const splittedFileName = fileName.split('.');
-  const fileExtension = splittedFileName[1];
-  const truncatedFileName = splittedFileName[0];
-  if (truncatedFileName.length >= 30) {
-    return `${truncatedFileName}... .${fileExtension ?? ''}`;
-  }
-  return fileName;
-};
-
-/**
- * Check if a mime type matches the set given in accept
- *
- * @link https://stackoverflow.com/a/66489392
- *
- * @param type the mime type to test, ex image/png
- * @param accept the mime types to accept, ex audio/*,video/*,image/png
- * @returns true if the mime is accepted, false otherwise
- */
-export function verifyAccept(type: string, accept?: Array<string>): boolean {
-  if (!accept) {
-    return true;
-  }
-  return accept.includes(type) || accept.includes(type.split('/')[0] + '/*');
-}
-
-/**
- * Returns the current status of the SingleFileInput based on his current state
+ * Returns the current status of the MISingleFileInput based on his current state
  *
  * @param file the current file
  * @param isLoading if the component is currently loading
  * @param isFileRejected if the file is rejected
- * @returns the current SingleFileInput status
+ * @returns the current MISingleFileInput status
  */
 export function getStatus(
   file: File | null,
@@ -89,7 +56,7 @@ export function getStatus(
 
 /**
  *
- * @param status the current status of the SingleFileInput
+ * @param status the current status of the MISingleFileInput
  * @returns the associated container styles
  */
 export function getContainerStyle(status: UploadStatus): SxProps {
@@ -134,16 +101,4 @@ export function getContainerStyle(status: UploadStatus): SxProps {
     default:
       return {};
   }
-}
-
-export function generateRandomID(): string {
-  /* eslint-disable no-bitwise */
-  return 'xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx'.replace(/[xy]/g, function (c) {
-    const randomNum =
-      typeof window !== 'undefined'
-        ? window.crypto.getRandomValues(new Uint8Array(1))[0] | 0
-        : (Math.random() * 16) | 0;
-    const v = c === 'x' ? randomNum : (randomNum & 0x3) | 0x8;
-    return v.toString(16);
-  });
 }
