+++ src/components/index.ts
@@ -45,3 +45,4 @@ export * from './MIPaper';
 export * from './MIBoxedModule';
 export * from './MIStepper';
 export * from './MIBreadcrumbs';
+export * from './MISingleFileInput';
