+++ src/types/singleFileInput.ts
@@ -1,0 +1,8 @@
+export enum UploadStatus {
+  IDLE = 'IDLE',
+  DRAG_OVER = 'DRAG_OVER',
+  LOADING = 'LOADING',
+  REJECTED = 'REJECTED',
+  ERROR = 'ERROR',
+  SELECTED = 'SELECTED',
+}
